requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic for settings management
pydantic==2.5.3
pydantic-settings==2.1.0

# JWT authentication
PyJWT==2.8.0
cryptography==41.0.7

# HTTP client for PocketBase
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0  # For TestClient
